+++ projects/fhi-angular-components/src/lib/fhi-angular-components.module.ts
@@ -9,6 +9,7 @@ import { FhiTableModule } from './fhi-table/fhi-table.module';
 import { FhiTreeViewNavigationModule } from './fhi-tree-view-navigation/fhi-tree-view-navigation.module';
 import { FhiTreeViewSelectionModule } from './fhi-tree-view-selection/fhi-tree-view-selection.module';
 
+import { FhiMonthSelectorComponent } from './fhi-time/fhi-month-selector/fhi-month-selector.component';
 import { FhiWeekpickerComponent } from './fhi-time/fhi-weekpicker/fhi-weekpicker.component';
 import { FhiWeekRangeComponent } from './fhi-time/fhi-week-range/fhi-week-range.component';
 import { FhiYearRangeComponent } from './fhi-time/fhi-year-range/fhi-year-range.component';
@@ -19,6 +20,7 @@ const FHI_ANGULAR_MODULES_AND_COMPONENTS = [
   FhiDateAndTimeModule,
   FhiDatepickerModule,
   FhiDateRangeComponent,
+  FhiMonthSelectorComponent,
   FhiMultiselectModule,
   FhiTableModule,
   FhiTreeViewNavigationModule,