+++ src/MOCK_DB_DATA/library-items/time-selectors/time-selector-month.ts
@@ -1,0 +1,34 @@
+import { LibraryItem, LibraryItemType } from 'src/app/views/shared/models/library-item.model';
+import { LibraryItemsSharedData as ITEMS } from '../library-items-shared-data';
+
+export const TimeSelectorMonth: LibraryItem[] = [{
+  id: ITEMS.TimeSelectorMonth.id,
+  title: ITEMS.TimeSelectorMonth.title,
+  type: LibraryItemType.angular,
+  exampleHtml: getExampleHtml(),
+  codeHtml: getCodeHtml(),
+  documentationHtml: getDocumentationHtml()
+}];
+
+
+/*
+ * Return value is ignored if LibraryItemType is not html
+ */
+function getExampleHtml(): string {
+  return ``;
+}
+  
+/*
+ * Return empty string to use a copy of exampleHtml as codeHtml.
+ * Return null to remove Code from library-item.
+ */
+function getCodeHtml(): string | null {
+  return `<fhi-month-selector></fhi-month-selector>`;
+}
+
+/*
+ * Return null to remove Documentation from library-item.
+ */
+function getDocumentationHtml(): string | null {
+  return null;
+}